Refactor some symplectic array-related functions

+++ src/qibochem/measurement/optimization.py
@@ -126,7 +126,7 @@ def _gc_measurement_mapping(expression: Expr, nqubits: int, method: str) -> tupl
     # Otherwise, expression is a sum of terms
     term_list = [_term_to_string(term) for term in expression.args if _term_to_string(term)[0] in ("X", "Y", "Z")]
     term_qubits = sorted({_get_qubit(op) for term in term_list for op in term.split()})
-    v_subspace = np.array([_pauli_to_symplectic(terms.split(), nqubits) for terms in term_list])
+    v_subspace = np.array([_pauli_to_symplectic(terms.split(), nqubits) for terms in term_list], dtype=np.uint8)
     v_basis = _binary_gaussian_elimination(v_subspace)
 
     dim_v = v_basis.shape[0]
